add gov rewards to tvl and manage rewards
