+++ src/app/pages/gov/gov.component.ts
@@ -8,11 +8,8 @@ import { InfoService } from '../../services/info.service';
 import { TerrajsService } from '../../services/terrajs.service';
 import { GoogleAnalyticsService } from 'ngx-google-analytics';
 import { GovPoolDetail } from './gov-pool/gov-pool.component';
-import { div, gt, minus, plus, times } from '../../libs/math';
 import { BalanceResponse } from '../../services/api/gov/balance_response';
 import { StateInfo } from '../../services/api/gov/state_info';
-import { AnchorMarketService } from '../../services/api/anchor-market.service';
-import { HttpClient } from '@angular/common/http';
 
 const LIMIT = 10;
 
@@ -38,16 +35,14 @@ export class GovComponent implements OnInit, OnDestroy {
     private gov: GovService,
     public info: InfoService,
     public terrajs: TerrajsService,
-    private anchorMarket: AnchorMarketService,
-    private httpClient: HttpClient,
     protected $gaService: GoogleAnalyticsService
   ) { }
 
   ngOnInit() {
     this.$gaService.event('VIEW_GOV_PAGE');
     this.connected = this.terrajs.connected
-      .subscribe(async _ => {
-        this.fetchPoolDetails();
+      .subscribe(async connected => {
+        this.info.initializeVaultData(connected);
         this.gov.config()
           .then(it => this.config = it);
         this.pollReset();
@@ -80,98 +75,6 @@ export class GovComponent implements OnInit, OnDestroy {
     this.polls.push(...res.polls);
     this.hasMore = res.polls.length >= LIMIT;
     this.$gaService.event('VIEW_GOV_PAGE', 'LOAD_MORE_POLL');
-  }
-
-  async fetchPoolDetails() {
-    const [state, rates] = await Promise.all([
-      this.anchorMarket.query({ epoch_state: {} }),
-      this.httpClient.get<any>(this.terrajs.settings.anchorAPI + '/deposit-rate').toPromise().catch(_ => undefined),
-      this.gov.state().then(it => this.stateInfo = it),
-      this.terrajs.isConnected
-        ? this.gov.balance().then(it => this.myBalance = it)
-        : Promise.resolve(null as BalanceResponse),
-      this.info.retrieveCachedStat(),
-      this.info.refreshPool(),
-    ]);
-
-    const anchorRatePerBlock = rates?.[0]?.deposit_rate ?? '0.000000041729682765';
-    const anchorRatePerYear = times(anchorRatePerBlock, 4656810);
-
-    const vaultFeeByPools = {};
-    let lockedBalance = '0';
-
-    const vaultFeeSlice = this.info.stat.vaultFee / this.stateInfo.pool_weight;
-    for (let n = 0; n < this.stateInfo.pools.length; n++) {
-      const involvedPools = this.stateInfo.pools.slice(n);
-      const sumTotalBalance = involvedPools.reduce((sum, pool) => sum + +pool.total_balance, 0);
-      const pivotPool = involvedPools[0];
-
-      for (const pool of involvedPools) {
-        const poolVaultFee = +pool.total_balance / sumTotalBalance * vaultFeeSlice * pivotPool.weight;
-        vaultFeeByPools[pool.days] = (vaultFeeByPools[pool.days] || 0) + poolVaultFee;
-      }
-    }
-
-    if (this.myBalance) {
-      const mostLockedBalance = this.myBalance.locked_balance.reduce((c, [_, { balance }]) => Math.max(c, +balance), 0);
-      lockedBalance = div(mostLockedBalance, CONFIG.UNIT);
-    }
-
-    this.poolDetails = this.stateInfo.pools
-      .map((pool) => {
-        const balanceInfo = this.myBalance?.pools.find(p => p.days === pool.days);
-        const userBalance = div(balanceInfo?.balance ?? 0, CONFIG.UNIT);
-        const userAUst = div(balanceInfo?.pending_aust ?? 0, CONFIG.UNIT);
-        const userProfit = times(userAUst, state.exchange_rate);
-        const unlockAt = balanceInfo?.unlock ? new Date(balanceInfo.unlock * 1000) : null;
-        const poolTvl = +pool.total_balance * +this.info.specPrice;
-        const apr = vaultFeeByPools[pool.days] / poolTvl;
-        const austApr = +anchorRatePerYear;
-
-        return {
-          userBalance,
-          userAUst,
-          userProfit,
-          austApr,
-          unlockAt,
-          days: pool.days,
-          apr: apr + apr * +anchorRatePerYear / 2,
-          balance: div(pool.total_balance, CONFIG.UNIT),
-          userAvailableBalance: '0', // populate after this mapping
-          moveOptions: [], // populate after this mapping
-        };
-      })
-      .map((current, _, poolDetails) => {
-        const sumOtherPoolsBalance = poolDetails.filter(d => d.days !== current.days)
-          .reduce((sum, d) => plus(sum, d.userBalance), '0');
-
-        const unreservedBalance = minus(lockedBalance, sumOtherPoolsBalance);
-        const userAvailableBalance = gt(unreservedBalance, 0)
-          ? minus(current.userBalance, unreservedBalance)
-          : current.userBalance;
-
-        const moveOptions = poolDetails.filter(d => d.days > current.days)
-          .map(({ days, userBalance, unlockAt }) => ({ days, userBalance, unlockAt }));
-
-        return {
-          ...current,
-          userAvailableBalance,
-          moveOptions,
-        };
-      });
-
-    this.calculateAPR();
-  }
-
-  calculateAPR() {
-    let sumGovAPR = 0;
-    let totalStaked = 0;
-    for (const pool of this.poolDetails) {
-      sumGovAPR += +pool.userBalance * pool.apr;
-      totalStaked += +pool.userBalance;
-    }
-
-    this.stakedInGovAPR = sumGovAPR / totalStaked;
   }
 
   trackPoolDetails(_: unknown, poolDetail: GovPoolDetail) {
